Build an Streamlit App on top of FastAPI

+++ utils/expense_calculator.py
@@ -1,6 +1,6 @@
 class Calculator:
     @staticmethod
-    def multiply(a: int, b: int) -> int:
+    def multiply(a: float, b: float) -> float:
         """
         Multiply two integers.
